feat: refactor page context management and update navigation links
