+++ frontend/src/components/Layout.tsx
@@ -1,18 +1,12 @@
 import * as React from "react";
 import {Navigation} from "./navigation/Navigation.tsx";
-import {createContext, ReactNode, useContext, useState} from "react";
+import {ReactNode, useState} from "react";
 import {Footer} from "./Footer.tsx";
+import {PageContext} from "../hooks/PageContext.tsx";
 
 interface LayoutProps {
     children: React.ReactNode
 }
-
-interface PageContextType {
-    page: string;
-    setPage: (newVal: string) => void
-}
-
-const PageContext = createContext<PageContextType | undefined>(undefined)
 
 export const PageProvider = ({children}: { children: ReactNode }) => {
     const [page, setPage] = useState("home");
@@ -31,14 +25,6 @@ export function Layout({children}: LayoutProps) {
                 {children}
                 <Footer/>
             </div>
-
         </>
     )
-}
-export const usePageContext = () => {
-    const context = useContext(PageContext)
-    if (!context) {
-        throw new Error("usePageContext must be used within a PageProvider");
-    }
-    return context
 }
